Add files via upload
A script to play a game of bingo during that weekly presentation, which could have been a e-mail. Disclaimer, this is a fun project which you and your colleagues can use in a work meeting, to enjoy that call more. You are free to use this script, though please do not hold me accountable for any uncomfortable situation during your call.